internal: Add support for non-Backends transitive dependencies
Summary:
At the moment, transitive dependencies are only supported by passing in the
backendsImpl itself. For example:

```
b.foo = MakeFoo(&b)
```

One of the disadvantages of this approach is that in MakeFoo(), you don't know
if the dependency has been created yet, so you can't it in the constructor -
you need to store `b` and use it later.

This diff adds better support for transitive dependencies, so you can do
something like:

```
b.foo = MakeFoo(b.bar)
```

You can assume inside MakeFoo() that b.bar has been created, and you can use
it in the constructor.

Reviewers: vaughan

Reviewed By: vaughan

Differential Revision: https://phabricator.corp.luno.com/D86710