Remove previously generated files
Summary:
Task T111462

This change add code to remove previously generated files before generating new ones. It also puts the file removal and creation code into functions and makes the files names constants.

# Notes
One issue with doing it this way is that the files will have been removed if `Generate` fails. We could try and keep a backup of the files and recreate them, if it's a problem.
We could also add some logic to `Generate` to ignore the previously generated files during generation, but then we'd have to pass in file name info to `Generate`.

Reviewers: neil

Reviewed By: neil

Subscribers: neil

Tags: #pod_dx

Differential Revision: https://phabricator.corp.luno.com/D96502