Add -in and -out flags
weld currently assumes the weld spec is in the current directory, and that you want to write `backends_gen.go` and/or `weld_gen.go` to the current directory as well. These flags let you control where the weld spec is, and where you want to write the output files to.

For example, you can do something like this:

```go
//go:generate weld -verbose -in=../../../state -out=..
```

This will look for the weld spec in `../../../state` and write the output files to `..`.