Refactor to accept multiple packages as arguments
Summary: This MR speeds weld up for the case where you need to reweld multiple packages.

In addition to being able to call `weld` in the package you want to reweld, you can now provide a list of packages to reweld on the command line.

The `-in` and `-out` flags have been removed.

The output has been tidied up and some emojis have been added to make it a bit more interesting to look at.